Detect error in epicsStrPrintEscaped.
PR #29 introduced the use of epicsStrPrintEscaped so that line breaks in
strings would be handled correctly. Unfortunately, this introduced a bug
(#48) that would cause save attempts to fail when the string to be
written was empty.

This bug was fixed with 58cd9958157fb470de02d4b9d7d1e719e86c5ff5 and
ddc1314415b0a90a3e25baaa19bb7d9da332d40d, but this introduced an issue
where Autosave would assume success, even if epicsStrPrintEscaped failed
to write anything, not because the string was empty, but because fprintf
failed or some reason.

This commit reintroduces the check whether epicsStrPrintEscaped has
written anything, but it now correctly handles the case where
epicsStrPrintEscaped does not write anything because the string that it
is supposed to write is empty.